requirements.txt
langchain==0.1.0
chromadb==0.4.22
python-dotenv==1.0.0
pypdf==4.0.1
tiktoken==0.5.2
langchain-community
sentence-transformers 